AR-801: Renamed AuthScreenService and moved it under App\Security
