+++ src/Controller/AuthScreenController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller;
 
-use App\Service\AuthScreenService;
+use App\Security\ScreenAuthenticator;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -12,7 +12,7 @@ use Symfony\Component\HttpKernel\Attribute\AsController;
 class AuthScreenController extends AbstractController
 {
     public function __construct(
-        private AuthScreenService $authScreenService
+        private ScreenAuthenticator $authScreenService
     ) {
     }
